Cleanup index js file, add helper js file, remove hardcoded API key from index file
