+++ .gitignore
@@ -1,5 +1,7 @@
-venv/
+my-venv/
 .vscode
 concertFinder/db.sqlite3
 concertFinder/concerts/settings.py
 concertFinder/concertFinder/.DS_Store
+.env
+staticfiles/
