+++ concertFinder/concertFinder/static/concertFinder/index.js
@@ -1,28 +1,11 @@
-const mapboxAccessToken =
-  "pk.eyJ1IjoiYXNoYXJwMDIiLCJhIjoiY2p5b3EwMTJyMTdoajNtbG1jZTJsaHJvYSJ9.KacigdAtzleu4QeM-dx7XQ";
-let userLatitude = null;
-let userLongitude = null;
+import { convertTravelTime, sortConcertList } from "./helpers.js";
 
-let currentEventMarkers = [];
-let userLocation = null;
-async function handleSubmit(event) {
-  const artistName = document.querySelector("#artist-input").value;
-  event.preventDefault();
-  clearList();
-  const myRequest = new Request(`/api/events?artist_name=${artistName}`);
-  const response = await fetch(myRequest);
-  const events = await response.json();
-  populateMap(events);
-  handleListView(events);
-}
-
-function clearList() {
-  const list = document.querySelector(".concert-list");
-  while (list.firstChild) {
-    list.removeChild(list.firstChild);
-  }
-}
-
+/******************************************
+ *          Fetch User Location           *
+ *                                        *
+ ******************************************/
+let userLatitude;
+let userLongitude;
 const successCallback = (position) => {
   userLatitude = position.coords.latitude;
   userLongitude = position.coords.longitude;
@@ -33,9 +16,65 @@ const errorCallback = (error) => {
 };
 navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
 
+/*************************************************************
+ *  Fetch concert data + distance data from user's location  *
+ *                                                           *
+ *************************************************************/
+
 let submitBtn = document.querySelector("#submit-btn");
 submitBtn.addEventListener("click", handleSubmit);
-let map = null;
+
+async function handleSubmit(event) {
+  const artistName = document.querySelector("#artist-input").value;
+  event.preventDefault();
+  clearList();
+  clearMap();
+  const myRequest = new Request(`/api/events?artist_name=${artistName}`);
+  const response = await fetch(myRequest);
+  const events = await response.json();
+  populateMap(events);
+  handleListView(events);
+}
+
+/*
+ * This function fetches the distance data for every venue currently in the
+ * events list, then updates that venue object with the travel distance. An
+ * additional mapping (lngLatDistances) is used to map {long},{lat} to distances
+ * so that we only need to fetch the distance once per venue (in the case of multiple shows at one venue)
+ */
+async function getDistanceData(events) {
+  const lngLatDistances = new Map(); //mappings from event long,lat strings to distances response
+  const userLngLat = `${userLongitude},${userLatitude}`;
+  await Promise.all(
+    events.map(async (event) => {
+      const eventLngLat = `${event.venue.longitude},${event.venue.latitude}`;
+      if (lngLatDistances.has(eventLngLat) == true) {
+        // find distance and add it to curr event
+        event.venue.distance = lngLatDistances.get(eventLngLat);
+      } else {
+        let mapboxMatrixURL = `https://api.mapbox.com/directions-matrix/v1/mapbox/driving/${userLngLat};`;
+        mapboxMatrixURL += `${eventLngLat}?access_token=${mapboxAccessToken}`;
+        const response = await fetch(mapboxMatrixURL, { mode: "cors" });
+        const distances = await response.json();
+        let distanceInSeconds = distances["durations"][0][1];
+        event.venue.distance = distanceInSeconds;
+        event.venue.distanceReadable = convertTravelTime(distanceInSeconds);
+        lngLatDistances.set(eventLngLat, event.venue.distance);
+      }
+    })
+  );
+}
+
+/************************************************
+ *    Render + populate map with concert data   *
+ *                                              *
+ ************************************************/
+let map;
+let currentEventMarkers = [];
+
+let hiddenInput = document.querySelector("#MAPBOX_API_KEY");
+let mapboxAccessToken = hiddenInput.value;
+
 function drawMap() {
   mapboxgl.accessToken = mapboxAccessToken;
   map = new mapboxgl.Map({
@@ -49,12 +88,15 @@ function drawMap() {
     .addTo(map);
 }
 
-function populateMap(shows) {
+function clearMap() {
   // Remove previously placed markers
   for (const marker of currentEventMarkers) {
     marker.remove();
   }
-  currId = 0;
+}
+
+function populateMap(shows) {
+  let currId = 0;
   shows.forEach((show) => {
     // create the popup
     let popupHTML = `<h2 class="headliner popup-headliner">${show.headliner}</h3>`;
@@ -83,55 +125,28 @@ function populateMap(shows) {
     currId++;
   });
 }
-/*
- * This function fetches the distance data for every venue currently in the
- * events list, then updates that venue object with the travel distance. An
- * additional mapping (lngLatDistances) is used to map {long},{lat} to distances
- * so that we only need to fetch the distance once per venue (in the case of multiple shows at one venue)
- */
-async function getDistanceData(events) {
-  const lngLatDistances = new Map(); //mappings from event long,lat strings to distances response
-  const userLngLat = `${userLongitude},${userLatitude}`;
-  await Promise.all(
-    events.map(async (event) => {
-      const eventLngLat = `${event.venue.longitude},${event.venue.latitude}`;
-      if (lngLatDistances.has(eventLngLat) == true) {
-        // find distance and add it to curr event
-        event.venue.distance = lngLatDistances.get(eventLngLat);
-      } else {
-        let mapboxMatrixURL = `https://api.mapbox.com/directions-matrix/v1/mapbox/driving/${userLngLat};`;
-        mapboxMatrixURL += `${eventLngLat}?access_token=${mapboxAccessToken}`;
-        const response = await fetch(mapboxMatrixURL, { mode: "cors" });
-        const distances = await response.json();
-        distanceInSeconds = distances["durations"][0][1];
-        event.venue.distance = distanceInSeconds;
-        event.venue.distanceReadable = convertTravelTime(distanceInSeconds);
-        lngLatDistances.set(eventLngLat, event.venue.distance);
-      }
-    })
-  );
 
-  // sort events by distance
-  events.sort((a, b) => {
-    if (a.venue.distance == null) {
-      return 1;
-    } else if (b.venue.distance == null) {
-      return -1;
-    } else {
-      return a.venue.distance - b.venue.distance;
-    }
-  });
+/*************************************************************
+ *   Display concert list + handle list item interactivity   *
+ *                                                           *
+ *************************************************************/
+
+async function handleListView(events) {
+  if (userLongitude == null || userLatitude == null) {
+    console.log("Please allow location on browser");
+    return;
+  }
+  await getDistanceData(events);
+  sortConcertList(events);
+  populateList(events);
+  setHoverAndClickEventsForListItems();
 }
 
-function convertTravelTime(durationInSeconds) {
-  const hours = durationInSeconds / 3600;
-  const minutes = (hours % 1) * 60;
-  let timeString = "";
-  if (hours >= 1) {
-    timeString += `${Math.round(hours)} Hours `;
+function clearList() {
+  const list = document.querySelector(".concert-list");
+  while (list.firstChild) {
+    list.removeChild(list.firstChild);
   }
-  timeString += `${Math.round(minutes)} Minutes`;
-  return timeString;
 }
 
 function populateList(events) {
@@ -155,19 +170,6 @@ function populateList(events) {
   }
 }
 
-async function handleListView(events) {
-  if (userLongitude == null || userLatitude == null) {
-    console.log("Please allow location on browser");
-    return;
-  }
-  await getDistanceData(events);
-  populateList(events);
-  console.log(events);
-  setHoverAndClickEventsForListItems();
-}
-
-let showDelay = 300,
-  hideDelay = 300;
 let itemHoverOverTimer, itemHoverOutTimer;
 let currentHighlightedMarker;
 function setHoverAndClickEventsForListItems() {
@@ -235,17 +237,3 @@ function unHighlightMarker(listItem) {
   showMarker.style.width = "20px";
   showMarker.style.height = "20px";
 }
-
-/*
-1. Handle zoom to fit markers
-2. styling
-3. handle multiple shows at same venue
-4. Add list of events ordered by distance from current location
-5. Add hover effect + interactivity on list items
-6. Style list!
-7. Handle non-driveable venues
-8. Have multiple list view options: sort by travel time or sort by date!
-9. Make list view scrollable within container!
-10. Add loading bar
-11. Convert travel time to hours/minutes!
-*/
