+++ concertFinder/concertFinder/views.py
@@ -1,5 +1,11 @@
+import environ
+
 from django.shortcuts import render
+
+env = environ.Env()
+environ.Env.read_env()
 
 
 def index(request):
-    return render(request, "concertFinder/index.html")
+    context = {"MAPBOX_API_KEY": env("MAPBOX_API_KEY")}
+    return render(request, "concertFinder/index.html", context)
